requirements.txt
apex==0.1
fire==0.4.0
hypertools==0.8.0
matplotlib==3.3.0
mmcv==1.4.6
mmcv_full==1.3.2
numpy==1.19.1
Pillow==9.0.1
plotly==5.6.0
progress==1.6
scikit_learn==1.0.2
scipy==1.5.2
tensorboardX==2.5
torch==1.6.0+cu101
torchvision==0.12.0
tqdm==4.48.2
